move tiles back to their original package
