+++ gradle.properties
@@ -1,5 +1,5 @@
 group=org.hexworks.zircon
-version=2019.0.21-PREVIEW
+version=2019.0.22-PREVIEW
 
 org.gradle.parallel=true
 org.gradle.daemon=true
