+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/api/Tiles.kt
@@ -1,7 +1,7 @@
 package org.hexworks.zircon.api
 
 import org.hexworks.zircon.api.builder.data.TileBuilder
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.data.Tile
 import kotlin.jvm.JvmStatic
