+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/api/data/CharacterTile.kt
@@ -1,7 +1,6 @@
-package org.hexworks.zircon.api.data.tile
+package org.hexworks.zircon.api.data
 
 import org.hexworks.zircon.api.color.TileColor
-import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.graphics.StyleSet
 import org.hexworks.zircon.api.modifier.Modifier
