+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/api/data/ImageTile.kt
@@ -1,8 +1,7 @@
-package org.hexworks.zircon.api.data.tile
+package org.hexworks.zircon.api.data
 
 import org.hexworks.zircon.api.behavior.TilesetOverride
 import org.hexworks.zircon.api.color.TileColor
-import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.graphics.StyleSet
 import org.hexworks.zircon.api.modifier.Modifier
 import org.hexworks.zircon.api.resource.TilesetResource
