+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/api/data/Tile.kt
@@ -4,9 +4,6 @@ import org.hexworks.cobalt.datatypes.Maybe
 import org.hexworks.zircon.api.behavior.Cacheable
 import org.hexworks.zircon.api.behavior.Drawable
 import org.hexworks.zircon.api.color.TileColor
-import org.hexworks.zircon.api.data.tile.CharacterTile
-import org.hexworks.zircon.api.data.tile.GraphicalTile
-import org.hexworks.zircon.api.data.tile.ImageTile
 import org.hexworks.zircon.api.graphics.StyleSet
 import org.hexworks.zircon.api.modifier.Border
 import org.hexworks.zircon.api.modifier.Modifier
