+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/api/tileset/impl/CP437TileMetadataLoader.kt
@@ -1,6 +1,6 @@
 package org.hexworks.zircon.api.tileset.impl
 
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.tileset.TileMetadataLoader
 import org.hexworks.zircon.api.tileset.TextureMetadata
