+++ zircon.core/common/src/main/kotlin/org/hexworks/zircon/internal/tileset/impl/VirtualTileset.kt
@@ -1,7 +1,7 @@
 package org.hexworks.zircon.internal.tileset.impl
 
 import org.hexworks.cobalt.datatypes.factory.IdentifierFactory
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.data.Position
 import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.tileset.TileTexture
