+++ zircon.jvm.examples/src/main/kotlin/org/hexworks/zircon/examples/CustomTileExample.kt
@@ -3,7 +3,7 @@ package org.hexworks.zircon.examples
 import org.hexworks.zircon.api.Positions
 import org.hexworks.zircon.api.SwingApplications
 import org.hexworks.zircon.api.Tiles
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.graphics.Symbols
 import org.hexworks.zircon.api.tileset.impl.CP437TileMetadataLoader
