+++ zircon.jvm.libgdx/src/main/kotlin/org/hexworks/zircon/internal/tileset/LibgdxTileset.kt
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.*
 import com.badlogic.gdx.graphics.g2d.*
 import com.github.benmanes.caffeine.cache.Caffeine
 import org.hexworks.cobalt.datatypes.Identifier
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.data.Position
 import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.modifier.TileTransformModifier
