+++ zircon.jvm.swing/src/main/kotlin/org/hexworks/zircon/internal/tileset/Java2DCP437Tileset.kt
@@ -3,7 +3,7 @@ package org.hexworks.zircon.internal.tileset
 import com.github.benmanes.caffeine.cache.Caffeine
 import org.hexworks.cobalt.datatypes.Identifier
 import org.hexworks.cobalt.datatypes.factory.IdentifierFactory
-import org.hexworks.zircon.api.data.tile.CharacterTile
+import org.hexworks.zircon.api.data.CharacterTile
 import org.hexworks.zircon.api.data.Position
 import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.api.modifier.*
