+++ zircon.jvm.swing/src/main/kotlin/org/hexworks/zircon/internal/tileset/Java2DImageDictionaryTileset.kt
@@ -2,7 +2,7 @@ package org.hexworks.zircon.internal.tileset
 
 import com.github.benmanes.caffeine.cache.Caffeine
 import org.hexworks.cobalt.datatypes.factory.IdentifierFactory
-import org.hexworks.zircon.api.data.tile.ImageTile
+import org.hexworks.zircon.api.data.ImageTile
 import org.hexworks.zircon.api.data.Position
 import org.hexworks.zircon.api.data.Tile
 import org.hexworks.zircon.internal.resource.TileType.IMAGE_TILE
